Fleshing out the ISale, implementing the first type of sale: SalePrice, and implementing a recursive algorithm that will get the best price (not required, nor tested fully yet)

+++ Blatant Shopping/CartService.cs
@@ -9,15 +9,15 @@ namespace BlatantShopping
 {
 	class CartService
 	{
-		public Dictionary<String, decimal> GetCartFromFile(String filename)
+		public Dictionary<String, int> GetCartFromFile(String filename)
 		{
 			return GetCartFromString(File.ReadAllLines(filename));
 		}
 
 
-		public Dictionary<String, decimal> GetCartFromString(String[] products)
+		public Dictionary<String, int> GetCartFromString(String[] products)
 		{
-			Dictionary<String, decimal> shoppingCart = new Dictionary<String, decimal>();
+			Dictionary<String, int> shoppingCart = new Dictionary<String, int>();
 
 			foreach (var product in products)
 			{
